simplify h2oModelInput. use function import instead of global variable
