+++ src/h2oModelInput/createListControl/filterItems.js
