+++ src/h2oModelInput/createListControl/incrementSelectionCount.js
